requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
sqlalchemy>=2.0
google-genai>=1.0
python-telegram-bot>=21.0
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0
streamlit>=1.38
pandas>=2.0
playwright>=1.40
